2024/08/17
As I read through the problem, my intuition said this needed Dynamic Programming.
This intuition on itself is an improvement.
I was willing to give it a chance.
So I read through the Discussion section. Many arguing this is a hard problem.
As I don't have enough time and I can't "immediately" see a promising strategy, I decided I will be checking the Editorial and read through the comments there, so I can focus on learning something.