2024/08/18
Today I decided to cheat hehehe.
In the first few minutes, I did see the overall strategy of generating the "ugly numbers" by exponentiation of 2, 3 and 5.
I saw I could use the Math.min but I though that would lead to too many redundant calculations. So I thought there could be deterministic way of establishing the ordered list. Example: before increasing the exponent of 2 twice (2x2 = 4), increase the exponent of 3, and before increasing it on 2 and 3 (2x3 = 6), increase the exponent of 5.
Instead of dwelling too much on it, I decided to approach this problem as I would in "real life", since N is bounded by 1690. I built a list of sorted numbers in a reliable (and time consuming) way and then I just query a pre-calculated table (maybe a DB).
In the end, the faster solution was still a on demand list building, which surprised me a lot!
My solution seems to be consuming the least amount of working memory (and this could be improved even further by querying the disk).
Now I will read the Editorial so I can clearly understand the efficient algorithm.