2024/08/20
I spent too much time on this one.
Majorly because I correctly identified it as dynamic programming and after a brief analysis I saw an overall strategy to get there. So I "had to" try.
This code didn't pass most testcases because it lacks performance.
I'm not sure if the code has bugs but I would say it most likely does.
Despite introducing some naive memoization that made the code slightly more performant than the original, it still got stuck, and on a relatively small array (length = 40). The solution should work on length = 100, so my tree is exploding and I'm considering many repeated cases.
I may have missed a different performant approach.
For example, this code first builds a tree and then traverses the tree and calculates stuff from bottom to top. It's very likely the calculations can be made on the go, during the first phase.
Finally giving up and reading the Editorial.
I'm very tired but I enjoyed the battle.