2024/08/29
Analysing and coming up with lots of ideas was fast.
Coming up with a general strategy was also fast.
Implementation was super slow. Majorly because I'm on a less focused day, but also because my solution has many moving parts.
In the end I was able to submit something original, the performance wasn't that bad.
I'm curious to check other solutions (I bet I'll find an elegant one) and to read the Editorial, for increased clarity.

I think this was the first time I implemented Disjoint-Set with Path Compression not on the nodes but on "categories", in this case it was on the rows and columns.

I'm tired but I'm celebrating and this was super fun.