2024/09/07
Today was objectively a fight.
Not because the problem was difficult, it was me who slept poorly and couldn't focus or enjoy this otherwise funny puzzle.
I came up with several approaches but wasn't willing to put the effort in code.
In the end, I found a way to lazily make it work. It landed below the last decile, both in terms of performance and space.
Today it would have been wiser to give up earlier and read the Editorial. I still decided to not give up. But poor sleep makes poor decisions ':)