2024/09/18
Today was a very difficult day for me.
Regarding this problem, I know I have the right overall strategy, I know all the edge cases, and yet I'm just not being capable to focus and express through code. I am just trying to force the code to work. Not the right mindset. I need to sleep. Maybe tomorrow I'll come to this again and solve it out of spite xD