2024/09/25
After the lately series of Problems in the recurrent theme of Trie, I had the right intuition. Since the Problem is marked as hard, I decided to just check both hints immediately, which confirmed the intuition.
The implementation was straightforward. I decided to try and use HashMaps.
During implementation I got many ideas on possible improvements, both in time and space, but I was committed with getting a working initial version.
My intuition is telling me I could try a much simpler approach with one single HashMap, populated with at most 1 million keys (because of Problem's restrictions), where each key is a prefix of some word.
I want to try it and compare.
After that, I'm expecting to find some nice neat trick in the Editorial.