2024/09/29
This was a huge fight.
I had most ideas almost immediately and implemented it under 1 hour. However, I missed I needed to keep the not only the Min, but also the 2nd Min, the 3rd Min, etc.
When I realized that, I noticed a Priority Queue wouldn't be O(1), so I came up with the idea of Linked Sets just in a few minutes.
However, implementing it ended up being a real struggle, because I kept feeling a bit confused, and 2nd guessing what would go where.
I think the main mistake was not using pen and paper. Perhaps I felt it was straightforward. But it wasn't. Once I jotted down just a basic schema, it helped immensely.
This took me lots of time but I just loved it. Real mental fight. Feels good in the end.
I still want to check the Editorial and learn something.