2024/10/01
Took me more time than expected, due to the later realization of negative numbers.
This works, but it is bottom 5%, almost 3x slower than the fastest solutions. Not so great on memory either (likely because the map is left untrimmed).
I also didn't check the hints.
I may be missing some trick.
So let's check the Editorial and learn something.