2024/10/11
I'm surprised this implementation was one of the most performants Time-wise.
It took me longer than desirable but it felt really good going through the process.
I felt this problem was one of those sitting just in the boundaries of my current ability (if we constraint the time under a couple of hours) and I could experiment some flow, going through the details of the problem in my mind, fighting with the questions that were arising as I felt getting closer and closer to the solution.
Such a good experience.
Note: I borrowed the MinHeap implementation. I could "eventually" implement it myself, I did it a few times already, although it would still take lots of time, specially debugging some details I don't have memorized, and in a "real world" scenario I wouldn't try to reinvent the wheel anyway.
Reading the Editorial and calling it a day.