2024/10/18
This took me way way too much time.
It was fun until it stopped being fun and became a hassle.
However, when I submitted it, apparently it beats 100% in terms of speed, and apparently by a good margin. That felt great for a moment.
I did not implement the whole Queue by myself. 
Wrapping up by reading the Editorial.