2024/10/20
I'm very surprised this implementation beats 100% in terms of speed (in Javascript), apparently by a large margin.
This took me about 1 hour, but I certainly overcomplicated the process in my head, changing approaches at least a couple of times. I know I could have submitted something reasonable under 40 minutes, and pen and paper would have been useful to bring clarity.
I'm also surprised because I don't think this code is optimized.
I'm sure there's a way to "compress" the 't's and 'f's before pushing them into the stack because we know the operation beforehand.
And as we pop, it's not needed to evaluate all the elements in the list, we can stop when we find the first TRUE of FALSE if the current operation is OR or AND, respectively.
Reading the Editorial for some tricks.