2024/10/23
Took me around 40 minutes but I'm foggy. It got 100% by a very large margin (but this is still the beginning of the day). Space was also 85% which is nice.
Edit: chopped off almost half of the runtime by pushing and popping twice instead of using tuples/arrays. Nice!
Checking the Editorial for tricks.