2024/10/23
Initially I got the feeling this could be done with one pass but I wanted to get a working solution relatively fast.
After reading the titles of the approaches in the Editorial, one said "one pass".
Being assured it was possible, I went for it. It beats (so far) 100% speed and memory. Nice!
I have to enjoy the fun I can have with these easier mediums, I guess. But now I spent more than 1 hour.