2024/10/25
My first idea was sorting the string but I decided to get creative and try a one pass approach.
It was fun trying that, with good enough results, although I took more time than I wanted.
Then I spent some additional 10 min on the sorting idea. It got 100% but at this point I think the Leetcode guys are helping with those results. Was it 4 in a row? I don't know. And doesn't matter much, it's just a good sign that I'm getting good implementations.
Checking the Editorial for tricks.