2024/10/28
I didn't enjoy reading this Problem and was struggling with motivation.
I took a warm blanket and drink, meditation for some minutes, wandered around and returned.
After that it was relatively straightforward.
I used sort. Is there a way to avoid it? My gut feeling is 50/50.
But I'm almost sure I'll find some tricks in the Editorial.