2024/11/09
Today it was really hard for me to focus. I couldn't reach even closer to a state of flow.
However, for a while, it was fun coming up with ideas, analysing the sub-problems as they arised. 
But at some point it stopped being pleasurable and became a grind. I thought about giving up, but I'm glad I didn't. I think the code is relatively performant, although it looks a complete mess. It's going to stay like that today.
I need some rest. I'll read the Editorial for some tricks, I bet there are some.