2024/11/10
It was interesting. Today I wasn't focusing again, and maybe I'm a bit tired of this bit manipulation streak, but I enjoyed being creative.
I had the idea of keeping a frequency array but I wasn't sure if it was going to be enough, since parsing the array of size 30 is not much doing so N times could become a problem.
I checked the Hints, they didn't help much.
I scrolled through the Discussion section and someone was complaining about how to remove the elements from the back, basically wondering how to apply the sliding window on a OR accumulator, the problem that motivated me to considering using the array in the first place. Some comments reinforced it would work.
So I decided to continuing implement it, and if that still failed to pass all tests I would give up.
In the end it passed, cool.
I'm almost sure the Editorial will reveal serious tricks. Let's see.