2024/11/19
This was an interesting one.
I came up with a general solution in 10 minutes but after failing the first submission it became clear there was more nuance. After failing the second submission I realized I had to pay more attention on how to keep track of the repeated values.

In the end, this code solves it, but it is in the very lower spectrum in terms of performance (although some of the submissions - using the same code - are saying the space is between 5% and 95%, so lots of variance in this space measurement).

I'll find something interesting and elegant in the Editorial.