2024/11/20
Such an interesting Problem.

I implemented something I was convinced to be the solution in less then 30 minutes.
After submission and failing a test case and understanding why, I got this clear feeling of "ah, you got me!". Which means I realized my overall strategy was wrong and I didn't know exactly how to proceed.
The only alternative I had considered was, on each turn, making the binary choice of taking the left or right element. But keeping track of that would explode.

I considered giving up but decided not to and headed for the Topics section, and it surprised me to see "sliding window". I opened the Hints, which give the solution away. In my head I confirmed the idea and thought "this is a smart trick, I wish I could have thought about it".

This is like a "reversed sliding window" because we are sliding the intervals on the extremes of the array, not in the middle. This is quite cool!

Fortunately I could reuse some code from the previous attempt. In the end this Problem cost me some more time then I wanted to, but at least the result was 100% for runtime and 86% for memory, which I'll take gladly.

Wrapping up as always, and moving on.