2024/11/23
Unfortunately I looked into Topics, just to check if it said "simulation" but I ended up reading "two pointers". I was able to make it worse, because I read a few comments in Discussion, and someone just said "or you can shift elements to the right and then rotate".
This just gave everything away, including the challenge, and I implemented this in less than 30 min.
It's 77% Runtime, so it's not amazing. I checked the most performant submission and it was smart, and shows my "for let w" loops are redundant work.
To the Editorial looking for some trick.