2024/11/24
Funny problem, at first it looked to me like it would need some sort of Dynamic Programming, it was cool realizing how to simplify it.
Pen and paper have been useful.
This implementation took 30 min (with a couple of distractions), passed all tests at 1st attempt and "beats" 100% Runtime. Nice!
It was 10% in terms of Memory though, I wonder if I could improve it, I didn't take the time to focus on optimizations.
To the Editorial and beyond.