2024/11/28
This one is hard and I didn't have the time (nor the will tbh) to implement it.
I read the Problem. It looks to me like we should try to find a path from zero to end. It feels like A* could be used here?
If we can find the path, it's done.
If we can't, we must have a "wall" of obstacles.
We process obstacles, which may be organized in many groups.
We find the groups that are walls (there may be more than 1, hard stuff indeed).
Then we need to find the smallest of obstacles that open that wall.
I'll read the Editorial and learn something.