2024/11/30
It's frustrating to not be available to handle these 3 hard Problems related with graphs.
From the top of my head, in this case we "just" have to find a Eulerian or Hamiltonian path (can't remember which is which), and we know it exists for sure.
This means we are guaranteed to have N nodes or N-2 nodes with even degree of edges.
I think we just have to start somewhere and keep adding edges to the front, until it's not possible anymore. If we still have edges to add, just start adding them to the back.
Then I would need to think a bit further about doing this in some "smart" way, and pay attention to the constraints.
Anyway, I'm going to check the Editorial and move on.