2024/12/02
Too bad I deleted the inner while loop after starting typing it. I had the right feeling but decided to submit as it was to keep it under 10 minutes but in the end I failed the 1st submission.
I also decided to go for the 2 pointers instead of sentence.split(' ') just for a bit of a challenge.